Extract preferences alerts into new component
